feat: Separate helper functions and classes (#2)
