+++ simple_visualizer/src/simple_visualizer/simple_visualizer.py
@@ -1,21 +1,10 @@
-import json
-from datetime import datetime, date
 from pathlib import Path
-from typing import List, Dict, Any
-
 from jinja2 import Environment, FileSystemLoader
 
 from api.components import VisualizerPlugin
 from api.model.graph import Graph
 
-
-class DateTimeEncoder(json.JSONEncoder):
-    """Custom JSON encoder that handles datetime and date objects."""
-
-    def default(self, obj: Any) -> str:
-        if isinstance(obj, (datetime, date)):
-            return obj.isoformat()
-        return super().default(obj)
+from .util import convert_nodes_to_dict, convert_edges_to_dict, custom_tojson
 
 
 class SimpleVisualizer(VisualizerPlugin):
@@ -33,51 +22,14 @@ class SimpleVisualizer(VisualizerPlugin):
         """Configure Jinja2 environment with custom filters and template loader."""
         template_path = Path(__file__).parent / "templates"
 
-        self._environment = Environment(
-            loader=FileSystemLoader(str(template_path))
-        )
-
-        # Register custom JSON filter that handles datetime objects
-        self._environment.filters['tojson'] = self._custom_tojson
+        self._environment = Environment(loader=FileSystemLoader(str(template_path)))
+        self._environment.filters['tojson'] = custom_tojson
         self._template = self._environment.get_template(self.TEMPLATE_NAME)
 
-    def _custom_tojson(self, obj: Any, **kwargs) -> str:
-        """Custom JSON serialization filter for Jinja2 templates."""
-        return json.dumps(obj, cls=DateTimeEncoder, **kwargs)
-
-    def _convert_nodes_to_dict(self, nodes) -> List[Dict[str, Any]]:
-        """Convert node objects to dictionary representation."""
-        return [
-            {
-                'id': node.id,
-                'name': node.id,
-                **node.data
-            }
-            for node in nodes
-        ]
-
-    def _convert_edges_to_dict(self, edges) -> List[Dict[str, str]]:
-        """Convert edge objects to dictionary representation."""
-        return [
-            {
-                'source': edge.src.id,
-                'target': edge.dest.id
-            }
-            for edge in edges
-        ]
-
     def visualize(self, data: Graph) -> str:
-        """
-        Generate HTML visualization of the graph data.
-
-        Args:
-            data: Graph object containing nodes and edges
-
-        Returns:
-            Rendered HTML string
-        """
-        nodes_list = self._convert_nodes_to_dict(data.nodes)
-        edges_list = self._convert_edges_to_dict(data.edges)
+        """Generate HTML visualization of the graph data."""
+        nodes_list = convert_nodes_to_dict(data.nodes)
+        edges_list = convert_edges_to_dict(data.edges)
 
         return self._template.render(
             nodes=nodes_list,
@@ -87,9 +39,7 @@ class SimpleVisualizer(VisualizerPlugin):
         )
 
     def name(self) -> str:
-        """Return the human-readable name of the plugin."""
         return self.PLUGIN_NAME
 
     def identifier(self) -> str:
-        """Return the unique identifier for the plugin."""
         return self.PLUGIN_IDENTIFIER
